module_10/04_event_hub/requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database & ORM
psycopg2-binary==2.9.9

# Authentication & Permissions
djangorestframework-simplejwt==5.5.1
django-filter==23.5

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0

# Development
ipython==8.17.2
django-extensions==3.2.3
black==23.12.1
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0
